+++ projects/ontimize-web-ngx/src/lib/components/table/extensions/dialog/export/o-table-export-dialog.component.ts
@@ -30,6 +30,7 @@ export class OTableExportDialogComponent implements OnInit, OnDestroy {
   protected exportService: OntimizeExportService;
   protected translateService: OTranslateService;
   protected oTableExportButtonService: OTableExportButtonService;
+  protected visibleButtons: string[];
   private subscription: Subscription = new Subscription();
 
   constructor(
@@ -40,6 +41,10 @@ export class OTableExportDialogComponent implements OnInit, OnDestroy {
     this.dialogService = injector.get(DialogService);
     this.translateService = this.injector.get(OTranslateService);
     this.oTableExportButtonService = this.injector.get(OTableExportButtonService);
+
+    if (config && Util.isDefined(config.visibleButtons)) {
+      this.visibleButtons = Util.parseArray(config.visibleButtons.toLowerCase(), true);
+    }
   }
 
   ngOnInit() {
@@ -115,6 +120,10 @@ export class OTableExportDialogComponent implements OnInit, OnDestroy {
         });
       }
     });
+  }
+
+  isButtonVisible(btn: string): boolean {
+    return !this.visibleButtons || (this.visibleButtons.indexOf(btn) !== -1);
   }
 
   protected handleError(err): void {