+++ projects/ontimize-web-ngx/src/lib/components/table/extensions/header/table-menu/o-table-export-configuration.class.ts
@@ -7,5 +7,6 @@ export class OTableExportConfiguration {
   filter?: object;
   mode: string;
   entity: string;
+  visibleButtons: string;
   options?: any;
 }